Switch away from `JsonSubTypes` library
It turns out NJsonSchema (which is pulled in by NSwag) already has
a `JsonInheritanceConverter()` which does exactly what we need.

This drops an extra dependency, and exposes it correctly to the
generated API ui (which the frontend uses to auto-generate its
service clients).

The downside is that there doesn't seem to be a way of declaring
known sub types programatically with this, which means it's going to
have declarative (i.e. attributes). Thus, we'll have to duplicate
attributes between Bson and Json for this.