Use a mongo pipeline to build the word ranks map
This was the last thing that needed to iterate through all the subs
directly, so I should be able to delete the code to read that stuff
after this.

Also, LINQ flipped its shit when it came to the Aggregate() pipeline
defintion, so I had to define it manually using mongo's own query
language. It's not too bad, to be honest.